# backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sahi Jagah API
  description: Property marketplace API for Indian tier-2/tier-3 cities
  version: 1.0.0
  contact:
    name: Sahi Jagah Support
    phone: '+917093187420'
servers:
  - url: http://localhost:5000/api/v1
    description: Development server
  - url: https://api.sahijagah.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Property:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        price:
          type: number
        propertyType:
          type: string
          enum: [rent, sale]
        bedrooms:
          type: integer
        bathrooms:
          type: integer
        area:
          type: number
        city:
          type: string
        address:
          type: string
        location:
          type: object
          properties:
            lat:
              type: number
            lng:
              type: number
        amenities:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        videos:
          type: array
          items:
            type: string
        ownerId:
          type: string
        status:
          type: string
          enum: [active, inactive, rented, sold]
        isVerified:
          type: boolean
        viewCount:
          type: integer
        favoritesCount:
          type: integer
        inquiriesCount:
          type: integer
...
paths:
...
  /properties/{id}:
    get:
      tags:
        - Properties
      summary: Get property details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Property details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '404':
          description: Property not found

    put:
      tags:
        - Properties
      summary: Update property
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Property'
      responses:
        '200':
          description: Property updated successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Only property owner can update
        '404':
          description: Property not found

    delete:
      tags:
        - Properties
      summary: Delete property (soft delete)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Property deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Property not found
...